update types, components - add BMAC component
